+++ components/PageMeta.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Head from 'next/head';
 
 import { joinUrl } from './utils/utils';
@@ -8,9 +7,19 @@ interface PageMetaProps {
   title: string;
   description: string;
   path: string;
+  previewImage?: string;
+  webPageStructuredData?: object;
+  articleStructuredData?: object;
 }
 
-const PageMeta: React.FC<PageMetaProps> = ({ title, description, path }) => (
+const PageMeta: React.FC<PageMetaProps> = ({
+  title,
+  description,
+  path,
+  previewImage,
+  webPageStructuredData,
+  articleStructuredData,
+}) => (
   <Head>
     <meta name="viewport" content="width=device-width,initial-scale=1" key="viewport" />
 
@@ -20,6 +29,9 @@ const PageMeta: React.FC<PageMetaProps> = ({ title, description, path }) => (
     {/* og:tags */}
     <meta key="page-og-title" property="og:title" content={title} />
     <meta key="page-og-description" property="og:description" content={description} />
+    {previewImage && (
+      <meta key="page-og-image" property="og:image" content={`https:${previewImage}`} />
+    )}
     {process.env.CANONICAL_URL && (
       <meta
         key="page-og-url"
@@ -27,13 +39,34 @@ const PageMeta: React.FC<PageMetaProps> = ({ title, description, path }) => (
         content={joinUrl(process.env.CANONICAL_URL, path)}
       />
     )}
+
+    {/* Twitter card */}
+    <meta name="twitter:title" content={title} />
+    <meta name="twitter:description" content={description} />
+    {previewImage && (
+      <meta key="page-twitter-image" name="twitter:image" content={`https:${previewImage}`} />
+    )}
+
+    {/* Structured data */}
+    {webPageStructuredData && (
+      <script
+        type="application/ld+json"
+        key="structured-data-webpage"
+        dangerouslySetInnerHTML={{
+          __html: JSON.stringify(webPageStructuredData),
+        }}
+      />
+    )}
+    {articleStructuredData && (
+      <script
+        type="application/ld+json"
+        key="structured-data-article"
+        dangerouslySetInnerHTML={{
+          __html: JSON.stringify(articleStructuredData),
+        }}
+      />
+    )}
   </Head>
 );
-
-PageMeta.propTypes = {
-  title: PropTypes.string.isRequired,
-  description: PropTypes.string.isRequired,
-  path: PropTypes.string.isRequired,
-};
 
 export default PageMeta;