+++ typings/index.ts
@@ -2,4 +2,5 @@ export * from './contentful';
 export * from './Meta';
 export * from './Pages';
 export * from './Project';
+export * from './StructuredData';
 export * from './ui';