+++ CHANGELOG.md
@@ -39,6 +39,40 @@ for any bug fixes.
 in case of vulnerabilities.
 -->
 
+## [1.2.1] - 2019-12-28
+
+### Removed
+
+- Removed all PropTypes (see release [1.2.0])
+
+### Fixed
+
+- Allow contentful images and videos in img/media CSP rules
+- eslintrc indentation
+
+## [1.2.0] - 2019-12-28
+
+### Added
+
+- Structured Data snippets
+- Twitter card meta tags
+
+### Changed
+
+- Update VSCode settings
+- More robust, crossbrowser native video player
+- Lazy images start the download of a new image earlier
+- Social sharing image for each page, instead of a global one
+
+### Removed
+
+- Removed PropTypes eslint validation and any PropTypes in the project, as they are redundant with TypeScript
+
+### Fixed
+
+- Disable native scroll restoration
+- Custom Rich Text Renderer, with list styles and asset links
+
 ## [1.1.0] - 2019-12-14
 
 ### Added
@@ -66,7 +100,8 @@ in case of vulnerabilities.
 
 - Next.js template project MVP
 
-[Unreleased]: https://github.com/ciampo/_nextjs-template/compare/v1.1.0...HEAD
-<!-- [X.Y.Z]: https://github.com/ciampo/_nextjs-template/compare/v1.1.0...vX.Y.Z -->
+[Unreleased]: https://github.com/ciampo/_nextjs-template/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/ciampo/_nextjs-template/compare/v1.2.0...v1.2.1
+[1.2.0]: https://github.com/ciampo/_nextjs-template/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ciampo/_nextjs-template/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/ciampo/_nextjs-template/releases/tag/v1.0.0